+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.2.0] - 2020-12-05
 [0.2.0]: https://github.com/mhassan1/yarn-plugin-aws-codeartifact/compare/v0.1.0...v0.2.0
 
+- Add authorization hook support for `npmRegistries` and `npmScopes`
 - Add `migrate` command that migrates `yarn.lock` entries to refer to AWS CodeArtifact for relevant packages
 - Rename `info` command to `version`